requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
dj-database-url==0.5.0
Django==4.0.3
djangorestframework==3.13.1
fonttools==4.32.0
gunicorn==20.1.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
qrcode==7.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
